+++ include/terraingraph/device/Rock.parm.h
@@ -2,7 +2,7 @@
 #error "You must define PARAM_INFO macro before include this file"
 #endif
 
-PARAM_INFO(Width,  Int,   width,  m_width,  (1024))
-PARAM_INFO(Height, Int,   height, m_height, (1024))
+PARAM_INFO(Width,  size_t, width,  m_width,  (1024))
+PARAM_INFO(Height, size_t, height, m_height, (1024))
 
 PARAM_INFO(Seed, Int, seed, m_seed, (0))
