+++ include/terraingraph/device/Voronoi.parm.h
@@ -2,8 +2,8 @@
 #error "You must define PARAM_INFO macro before include this file"
 #endif
 
-PARAM_INFO(Width,  Int,   width,  m_width,  (1024))
-PARAM_INFO(Height, Int,   height, m_height, (1024))
+PARAM_INFO(Width,  size_t, width,  m_width,  (1024))
+PARAM_INFO(Height, size_t, height, m_height, (1024))
 
 PARAM_INFO(Displacement,   Float, displacement, m_displacement, (1.0f))
 PARAM_INFO(EnableDistance, Bool,  enable_dist,  m_enable_dist,  (false))
