+++ source/device/TurbulenceNoise2.cpp
@@ -62,8 +62,8 @@ void TurbulenceNoise2::Execute(const std::shared_ptr<dag::Context>& ctx)
     NoiseModule prev_mod(*prev_hf);
     noise.SetSourceModule(0, prev_mod);
 
-    for (int y = 0; y < h; ++y) {
-        for (int x = 0; x < w; ++x) {
+    for (size_t y = 0; y < h; ++y) {
+        for (size_t x = 0; x < w; ++x) {
             const float fx = static_cast<float>(x) / w;
             const float fy = static_cast<float>(y) / h;
             double v = noise.GetValue(fx, 0, fy);
